add AddUser component and route
